pyproject.toml
[tool.poetry]
name = "general_dagster_mlflow"
version = "0.1.0"
description = ""
authors = ["Heiko Kulinna <Heikokulinna@gmx.de>"]


# to update only poetry lock file
# poetry update --lock -vvv

[tool.poetry.dependencies]
python = ">=3.8.4, <3.9"
pandas = "^1.3.4"
numpy = "^1.21.3"
plotly = "^5.3.1"

mlflow = "==1.14.1"    #"==1.20.1"   # 1.14.1
azure-storage-blob = "==12.7.1"
azure-identity ="^1.7.0"

# make it work again
protobuf = "< 3.20.0"
jinja2 = "< 2.11"
MarkupSafe = "< 2.1"
itsdangerous = "< 2.1"
ipython_genutils = "*"   # 0.2.0

PyYAML = "^6.0"
pyarrow = "^6.0.0"
sklearn = "*"

dagit = "0.13.5"
dagster = "0.13.5"
dagster-docker = "0.13.5"
dagster-postgres = "0.13.5"
dagster-k8s = "0.13.5"

python-dotenv = "^0.19.2"
"yaml-1.3" = ">= 0.1.0"


[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
isort = "^5.8.0"
pre-commit = "^2.12.1"
pylint = "^2.12.2"



[tool.black]
line-length = 79


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

